# docs/swagger.json
{"openapi":"3.0.0","info":{"description":"","version":"1.0.0","title":"Image Storage API's"},"host":"localhost:8080","schemes":["http"],"basePath":"/","securityDefinitions":{},"paths":{"/add/album":{"post":{"parameters":[],"requestBody":{"content":{"application/json":{"schema":{"properties":{"album_tittle":{"type":"string"}},"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"properties":{"album_tittle":{"example":"test","type":"string"},"message":{"example":"Album created sucessfully","type":"string"}},"type":"object"}}},"description":"Success"},"400":{"content":{"application/json":{"schema":{"properties":{"error":{"example":"album already exist","type":"string"}},"type":"object"}}},"description":"Failed"}},"security":[],"summary":"Create Album","tags":["Album"]}},"/add/image":{"post":{"parameters":[],"requestBody":{"content":{"multipart/form-data":{"schema":{"properties":{"album_tittle":{"type":"string"},"image":{"type":"file"}},"type":"object"}}}},"responses":{"200":{"content":{"application/json":{"schema":{"properties":{"album_tittle":{"example":"test","type":"string"},"image_count":{"example":10,"type":"int"},"image_id":{"example":12,"type":"int"},"image_name":{"example":"xynvnd.png","type":"string"}},"type":"object"}}},"description":"Success"},"400":{"content":{"application/json":{"schema":{"properties":{"error":{"example":"image already exist","type":"string"}},"type":"object"}}},"description":"Failed"}},"security":[],"summary":"Add Image","tags":["Image"]}},"/del/album/{tittle}":{"delete":{"parameters":[{"in":"path","name":"tittle","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"properties":{"album_tittle":{"example":"test","type":"string"},"message":{"example":"deleted successfully","type":"string"}},"type":"object"}}},"description":"Success"},"400":{"content":{"application/json":{"schema":{"properties":{"error":{"example":"album doesn't exist","type":"string"}},"type":"object"}}},"description":"Failed"}},"security":[],"summary":"Remove Album","tags":["Album"]}},"/del/image/{id}":{"delete":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"properties":{"album_tittle":{"example":"test","type":"string"},"image_path":{"example":"/albums/swgr/xudgfjf.png","type":"string"},"message":{"example":"deleted successfully","type":"string"}},"type":"object"}}},"description":"Success"},"400":{"content":{"application/json":{"schema":{"properties":{"error":{"example":"image doesn't exist","type":"string"}},"type":"object"}}},"description":"Failed"}},"security":[],"summary":"Remove Image","tags":["Image"]}},"/get/image/{id}":{"get":{"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"image/jpeg":{"schema":{"type":"object"}}},"description":"Success"},"400":{"content":{"application/json":{"schema":{"properties":{"error":{"example":"Image id not found","type":"string"}},"type":"object"}}},"description":"Failed"}},"security":[],"summary":"Get Image","tags":["Image"]}}},"components":{"schemas":null}}